worker/requirements.txt
google-cloud-storage
google-cloud-pubsub
flask
pyodbc
moviepy
Pillow
requests